WinTerm/requirements.txt
psutil
termcolor
pyfiglet
googlesearch-python
phonenumbers
pypdf2